Basic Code Arm
Arm contains three motors and two servos. One motor is in charge of lifting the full system with the left and right bumpers (left down, right up). Another motor is in charge of rotating the entire system about a "shoulder" with the left and right triggers. The final motor rotates half of the system about an "elbow" with the left and right dpad buttons. The first servo rotates the pinching mechanism about a "wrist" with the A and X buttons. The second servo pinches with the Y and B buttons.